Рефакторит экран предложения. WIP: далее сделать один компонент с условием.
