+++ project/src/components/app/app.tsx
@@ -1,8 +1,7 @@
 import {Route, BrowserRouter, Routes} from 'react-router-dom';
 import {HelmetProvider} from 'react-helmet-async';
-import MainScreen from '../../pages/main-screen/main-screen';
 import NotFoundScreen from '../../pages/not-found-screen//not-found-screen';
-import OfferScreen from '../../pages/offer-screen/offer-screen';
+import OfferScreenNotLogged from '../../pages/offer-screen-not-logged/offer-screen-not-logged';
 import LoginScreen from '../../pages/login-screen/login-screen';
 import ScrollToTop from '../scroll-to-top/scroll-to-top';
 
@@ -10,6 +9,7 @@ import {AppRoute} from '../../const/app-route';
 
 import type { City} from '../../@types/offer-types';
 import { Review } from '../../@types/review-types';
+import MainScreenNotLogged from '../../pages/main-screen-not-logged/main-screen-not-logged';
 
 type AppProps = {
   cities: City[];
@@ -24,7 +24,7 @@ function App({cities, reviews}: AppProps): JSX.Element {
         <Routes>
           <Route
             path={AppRoute.Main}
-            element={<MainScreen cities = {cities}/>}
+            element={<MainScreenNotLogged cities = {cities}/>}
           />
           <Route
             path={AppRoute.Login}
@@ -33,7 +33,7 @@ function App({cities, reviews}: AppProps): JSX.Element {
           <Route
             path={AppRoute.Offer}
             element={
-              <OfferScreen
+              <OfferScreenNotLogged
                 reviews={reviews}
               />
             }
